Explicitly call accessibility tests because of flakiness
