+++ playwright/tests/office/servicescounseling/servicesCounselingTestFixture.js
@@ -13,13 +13,6 @@ export class ServiceCounselorPage extends OfficePage {
    */
   constructor(officePage) {
     super(officePage.page, officePage.request);
-    Object.entries(this.waitForPage).forEach(([key, value]) => {
-      // eslint-disable-next-line security/detect-object-injection
-      this.waitForPage[key] = async () => {
-        await value();
-        await this.runAccessibilityAudit();
-      };
-    });
   }
 
   waitForPage = {
