+++ playwright/tests/utils/waitForPage.js
@@ -1,5 +1,6 @@
 // @ts-check
 import * as base from '@playwright/test';
+import { checkA11y } from 'axe-playwright';
 
 export class WaitForPage {
   /**
@@ -13,137 +14,193 @@ export class WaitForPage {
   }
 
   /**
+   * Run an accessibility audit against the page in its current state
+   *
+   */
+
+  async runAccessibilityAudit() {
+    if (process.env.A11Y_AUDIT) {
+      await checkA11y(
+        this.page,
+        undefined,
+        {
+          detailedReport: true,
+        },
+        // skip failures
+        true,
+        'default',
+      );
+    }
+  }
+
+  /**
    * @returns {Promise<void>}
    */
   async localLogin() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Select an Existing User' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingConus() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Where are you moving?' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingDodId() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Create your profile' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingName() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Name' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingContactInfo() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Your contact info' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingDutyLocation() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Current duty location' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingCurrentAddress() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Current pickup address' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingBackupAddress() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Backup address' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async onboardingBackupContact() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { name: 'Backup contact' })).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async home() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByTestId('stepContainer1').getByText('Profile complete')).toBeVisible();
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async ordersDetails() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { level: 1 })).toHaveText('Tell us about your move orders');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async ordersUpload() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { level: 1 })).toHaveText('Upload your orders');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async aboutShipments() {
+    await this.runAccessibilityAudit();
     await base
       .expect(this.page.getByRole('heading', { level: 1 }))
       .toHaveText('Things to know about selecting shipments');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async selectShipmentType() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { level: 1 })).toHaveText('How should this shipment move?');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async hhgShipment() {
+    await this.runAccessibilityAudit();
     await base
       .expect(this.page.getByRole('heading', { level: 1 }))
       .toHaveText('Movers pack and transport this shipment');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async ntsShipment() {
+    await this.runAccessibilityAudit();
     await base
       .expect(this.page.getByRole('heading', { level: 1 }))
       .toHaveText('Where and when should the movers pick up your things going into storage?');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async ntsReleaseShipment() {
+    await this.runAccessibilityAudit();
     await base
       .expect(this.page.getByRole('heading', { level: 1 }))
       .toHaveText('Where and when should the movers deliver your things from storage?');
+    await this.runAccessibilityAudit();
   }
 
   /**
    * @returns {Promise<void>}
    */
   async reviewShipments() {
+    await this.runAccessibilityAudit();
     await base.expect(this.page.getByRole('heading', { level: 1 })).toHaveText('Review your details');
+    await this.runAccessibilityAudit();
   }
 }
